@@ -1,4 +1,4 @@
-class Belt < ActiveRecord::Base
+class Belt < ApplicationRecord
   has_many :members
   default_scope { order("order_by ASC") }
 end
