@@ -1,4 +1,4 @@
-class CreateMeetingTypes < ActiveRecord::Migration
+class CreateMeetingTypes < ActiveRecord::Migration[4.2]
   def change
     create_table :meeting_types do |t|
       t.string :name
