@@ -1,4 +1,4 @@
-class AddSlugToSchools < ActiveRecord::Migration
+class AddSlugToSchools < ActiveRecord::Migration[4.2]
   def change
     add_column :schools, 'slug', :string
   end
