@@ -19,7 +19,5 @@
 //= require members
 //= require assistants
 //= require jquery_blockui
-//= require react
-//= require react_ujs
-//= require components
+//= require server_rendering
 //= require_tree .
