@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,88 +10,85 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20190610190841) do
+ActiveRecord::Schema.define(version: 2019_06_10_190841) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "belts", force: :cascade do |t|
-    t.string   "name"
-    t.text     "comment"
-    t.boolean  "is_active"
-    t.integer  "order_by"
+  create_table "belts", id: :serial, force: :cascade do |t|
+    t.string "name"
+    t.text "comment"
+    t.boolean "is_active"
+    t.integer "order_by"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  create_table "meeting_members", force: :cascade do |t|
-    t.integer  "meeting_id"
-    t.integer  "member_id"
-    t.integer  "role_id"
-    t.integer  "belt_id"
+  create_table "meeting_members", id: :serial, force: :cascade do |t|
+    t.integer "meeting_id"
+    t.integer "member_id"
+    t.integer "role_id"
+    t.integer "belt_id"
+    t.datetime "created_at"
+    t.datetime "updated_at"
+    t.index ["belt_id"], name: "index_meeting_members_on_belt_id"
+    t.index ["meeting_id"], name: "index_meeting_members_on_meeting_id"
+    t.index ["member_id", "meeting_id", "role_id"], name: "index_meeting_members_on_member_id_and_meeting_id_and_role_id", unique: true
+    t.index ["member_id"], name: "index_meeting_members_on_member_id"
+    t.index ["role_id"], name: "index_meeting_members_on_role_id"
+  end
+
+  create_table "meeting_types", id: :serial, force: :cascade do |t|
+    t.string "name"
+    t.text "comment"
+    t.boolean "is_active"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  add_index "meeting_members", ["belt_id"], name: "index_meeting_members_on_belt_id", using: :btree
-  add_index "meeting_members", ["meeting_id"], name: "index_meeting_members_on_meeting_id", using: :btree
-  add_index "meeting_members", ["member_id", "meeting_id", "role_id"], name: "index_meeting_members_on_member_id_and_meeting_id_and_role_id", unique: true, using: :btree
-  add_index "meeting_members", ["member_id"], name: "index_meeting_members_on_member_id", using: :btree
-  add_index "meeting_members", ["role_id"], name: "index_meeting_members_on_role_id", using: :btree
-
-  create_table "meeting_types", force: :cascade do |t|
-    t.string   "name"
-    t.text     "comment"
-    t.boolean  "is_active"
-    t.datetime "created_at"
-    t.datetime "updated_at"
-  end
-
-  create_table "meetings", force: :cascade do |t|
-    t.integer  "meeting_type_id"
+  create_table "meetings", id: :serial, force: :cascade do |t|
+    t.integer "meeting_type_id"
     t.datetime "met"
-    t.text     "comment"
+    t.text "comment"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.integer  "school_id",       default: 1
+    t.integer "school_id", default: 1
+    t.index ["meeting_type_id"], name: "index_meetings_on_meeting_type_id"
+    t.index ["met"], name: "index_meetings_on_met"
   end
 
-  add_index "meetings", ["meeting_type_id"], name: "index_meetings_on_meeting_type_id", using: :btree
-  add_index "meetings", ["met"], name: "index_meetings_on_met", using: :btree
-
-  create_table "members", force: :cascade do |t|
-    t.string   "first_name"
-    t.string   "last_name"
-    t.integer  "belt_id"
-    t.integer  "school_id"
-    t.text     "comment"
-    t.boolean  "is_active"
-    t.boolean  "is_teacher"
-    t.boolean  "is_kid"
+  create_table "members", id: :serial, force: :cascade do |t|
+    t.string "first_name"
+    t.string "last_name"
+    t.integer "belt_id"
+    t.integer "school_id"
+    t.text "comment"
+    t.boolean "is_active"
+    t.boolean "is_teacher"
+    t.boolean "is_kid"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.boolean  "is_quark",   default: false
-    t.boolean  "is_teen",    default: false
+    t.boolean "is_quark", default: false
+    t.boolean "is_teen", default: false
+    t.index ["belt_id"], name: "index_members_on_belt_id"
+    t.index ["school_id"], name: "index_members_on_school_id"
   end
 
-  add_index "members", ["belt_id"], name: "index_members_on_belt_id", using: :btree
-  add_index "members", ["school_id"], name: "index_members_on_school_id", using: :btree
-
-  create_table "roles", force: :cascade do |t|
-    t.string   "name"
-    t.text     "comment"
-    t.boolean  "is_active"
+  create_table "roles", id: :serial, force: :cascade do |t|
+    t.string "name"
+    t.text "comment"
+    t.boolean "is_active"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  create_table "schools", force: :cascade do |t|
-    t.string   "name"
-    t.text     "comment"
-    t.boolean  "is_active"
+  create_table "schools", id: :serial, force: :cascade do |t|
+    t.string "name"
+    t.text "comment"
+    t.boolean "is_active"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.string   "slug"
+    t.string "slug"
   end
 
 end
